requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
alembic
pandas
psycopg2-binary
python-dotenv
poetry 